services/trainer/requirements.txt
mlflow==2.13.1
prefect==2.16.5
google-cloud-storage==2.16.0
pydantic==2.6.4
numpy==1.26.4
requests==2.31.0
